+++ odoo/songs/install/data_demo.py
@@ -6,7 +6,7 @@ from pkg_resources import resource_stream
 
 import anthem
 from anthem.lyrics.loaders import load_csv_stream
-from ..common import req, load_users_csv
+from ..common import req
 
 """ File for demo data
 
@@ -21,34 +21,6 @@ def import_customers(ctx):
     """ Importing customers from csv """
     content = resource_stream(req, 'data/demo/customers.csv')
     load_csv_stream(ctx, 'res.partner', content, delimiter=',')
-
-
-@anthem.log
-def import_employee(ctx):
-    """ Importing employee from csv """
-    content = resource_stream(req, 'data/demo/hr.employee.csv')
-    load_csv_stream(ctx, 'hr.employee', content, delimiter=',')
-
-
-@anthem.log
-def import_partner_contact(ctx):
-    """ Importing partner_contact from csv """
-    content = resource_stream(req, 'data/demo/res.partner.contact.csv')
-    load_csv_stream(ctx, 'res.partner', content, delimiter=',')
-
-
-@anthem.log
-def import_partner(ctx):
-    """ Importing partner from csv """
-    content = resource_stream(req, 'data/demo/res.partner.csv')
-    load_csv_stream(ctx, 'res.partner', content, delimiter=',')
-
-
-@anthem.log
-def import_users(ctx):
-    """ Importing users from csv """
-    content = resource_stream(req, 'data/demo/res.users.csv')
-    load_users_csv(ctx, content, delimiter=',')
 
 
 @anthem.log
@@ -68,9 +40,5 @@ def import_project_tasks(ctx):
 @anthem.log
 def main(ctx):
     """ Loading demo data """
-    import_users(ctx)
-    import_partner(ctx)
-    import_partner_contact(ctx)
-    import_employee(ctx)
     import_projects(ctx)
     import_project_tasks(ctx)